+++ source/judge.cpp
@@ -13,7 +13,9 @@ class Command_judge: public App
         options.addOption(Poco::Util::Option("time","t","change time limit").argument("time",true));
         options.addOption(Poco::Util::Option("isuf","is","specify input file suf").argument("suf",true));
         options.addOption(Poco::Util::Option("osuf","os","specify output file suf").argument("suf",true));
+        options.addOption(Poco::Util::Option("ofile","of","specify std").argument("file",true));
         options.addOption(Poco::Util::Option("multithread","mul","turn on multithreading").noArgument());
+        options.addOption(Poco::Util::Option("dorecompile","dore","force recompile").noArgument());
         App::defineOptions(options);
     }
     void displayHelp(Poco::Util::HelpFormatter *helpFormatter)
@@ -34,6 +36,7 @@ class Command_judge: public App
         
         const std::string _ans_name="ans";
         const std::string _chk_name="checker";
+        const std::string _out_name="std";
         // init name
         fil ans=[&]()
         {
@@ -47,8 +50,10 @@ class Command_judge: public App
             add_file(_run_chk,chk_str);
             return add_namesuf(get_file(chk_str),"cpp");
         }();
-        // init time
+        fil out=check_option("ofile")?add_namesuf(get_file(get_option("ofile")),"cpp"):fil();
+        // init config
         if(check_option("time")) change_time_limit((tim)std::stoi(get_option("time")));
+        if(check_option("dorecompile")) if_skip_compiled=false;
         // init data
         if(args.size()==0)
         {
@@ -60,11 +65,13 @@ class Command_judge: public App
         // find file
         if(ans==fil()||!ans.exists()) {print_result(_ans_name,res::type::NF);return EXIT_NOINPUT;}
         if(chk==fil()||!chk.exists()) {print_result(_chk_name,res::type::NF);return EXIT_NOINPUT;}
+        if(out!=fil()&&!out.exists()) {print_result(_out_name,res::type::NF);return EXIT_NOINPUT;}
         if(show_file_info) scout<<termcolor::bright_grey<<print_type({"","","\n"},{{_ans_name+": ",ans},{_chk_name+": ",chk}},true)<<ANSI::move_up*2<<termcolor::reset;
         // compile file
         printer *print=new printer({"Compiling.","Compiling..","Compiling..."},(tim)150);print->start();
         th_compiler *run_compiler=new th_compiler();
         run_compiler->add({{_ans_name,ans},{_chk_name,chk}},data_compile_argu);
+        if(out!=fil()) run_compiler->add(_out_name,out,data_compile_argu);
         run_compiler->wait_all();
         {
             std::string name=run_compiler->get_all();
@@ -112,7 +119,8 @@ class Command_judge: public App
         // judge
         if(check_option("multithread"))
         {
-            unsigned ac_sum=0,runned_sum=0,add_sum=0,get_sum=0;
+            ssetenv(".data.chk_outputs","\"off\"");
+            unsigned ac_sum=0,runned_sum=0,add_sum=0,output_sum=0;
             th_judger run_judger;
             std::vector<std::string> judge_list;
             std::map<std::string,unsigned> num_list;
@@ -122,10 +130,11 @@ class Command_judge: public App
                 if(datas[name]["in"].is_null()&&datas[name]["out"].is_null()) return;
                 fil run_dir=default_data_dir/"tmp_data"/std::to_string(num_list[name]);
                 run_dir.createDirectory();
-                fil in_file=system_nul,out_file=system_nul,ans_file=run_dir/"data.ans",chk_file=run_dir/"data.txt";
+                fil in_file=system_nul,out_file=system_nul,ans_file=run_dir/"data.ans",chk_file=run_dir/"data.txt",tmp_out=fil();
                 if(!datas[name]["in"].is_null()) ((fil)(std::string)datas[name]["in"]).copyTo((in_file=run_dir/"data.in").path());
                 if(!datas[name]["out"].is_null()) ((fil)(std::string)datas[name]["out"]).copyTo((out_file=run_dir/"data.out").path());
-                run_judger.add(name,(new judger(ans,chk,in_file,out_file,ans_file,chk_file)));
+                else if(out!=fil()) out_file=run_dir/"data.out",tmp_out=out;
+                run_judger.add(name,(new judger(ans,chk,in_file,out_file,ans_file,chk_file))->set_out(tmp_out));
             };
             while(add_sum<max_thread_num&&add_sum<judge_list.size()) add(judge_list[add_sum++]);
             std::string name;
@@ -153,16 +162,24 @@ class Command_judge: public App
                 output_chk_file
                 <<"    infile: "<<add_squo(get_file(!datas[name]["in"].is_null()?(std::string)datas[name]["in"]:system_nul))
                 <<", outfile: "<<add_squo(get_file(!datas[name]["out"].is_null()?(std::string)datas[name]["out"]:system_nul))<<"\n";
+                if(target->out!=fil()) output_chk_file<<"    "+_out_name+": "<<add_squo(target->out)<<"\n";
                 output_chk_file<<"    result: "<<target->result<<"\n";
                 output_chk_file<<print_type({"    "," time: "," exit_code: ","\n"},{
                     {_ans_name+":",target->time,target->exit_code},
-                    {_chk_name+":",target->chk_runner->time,target->chk_runner->exit_code}});
+                    {_chk_name+":",target->chk_runner->time,target->chk_runner->exit_code},
+                    {_out_name+":",target->out_runner?target->out_runner->time:(tim)-1,target->out_runner?target->out_runner->exit_code:-1}});
                 output_chk_file<<std::string("*")*50;
                 output_chk_file.close();
                 if(target->result.istrue()) ++ac_sum;
                 if(target->result.isfalse())
                 {
-                    run_dir.copyTo((default_data_dir/(data_name+" - "+get_short_resultname(target->result))).path());
+                    ++output_sum;
+                    pat target_dir=get_file(replace_env(data_file_str,running_path,env_args::data(data_name,target->result,"",runned_sum,output_sum))).path();
+                    ((fil)target_dir.parent()).createDirectories();
+                    if((run_dir/"data.in").exists()) (run_dir/"data.in").copyTo(replace_extension(target_dir,"in").toString());
+                    if((run_dir/"data.out").exists()) (run_dir/"data.out").copyTo(replace_extension(target_dir,"out").toString());
+                    if((run_dir/"data.ans").exists()) (run_dir/"data.ans").copyTo(replace_extension(target_dir,"ans").toString());
+                    if((run_dir/"data.txt").exists()) (run_dir/"data.txt").copyTo(replace_extension(target_dir,"txt").toString());
                 }
                 #undef run_dir
             }
@@ -170,9 +187,11 @@ class Command_judge: public App
         }
         else
         {
-            unsigned runned_sum=0,ac_sum=0;
+            unsigned runned_sum=0,ac_sum=0,output_sum=0;
             for(auto i:datas.items())
             {
+                ssetenv("runned_sum",std::to_string(runned_sum));
+                ssetenv("outputed_sum",std::to_string(output_sum));
                 // init file
                 if(i.value()["in"].is_null()&&i.value()["out"].is_null()) continue;
                 const std::string data_name=[&]()
@@ -190,11 +209,12 @@ class Command_judge: public App
                 ++runned_sum;
                 fil run_dir=default_data_dir/"datas"/data_name;
                 run_dir.createDirectory();
-                fil in_file=system_nul,out_file=system_nul,ans_file=run_dir/"data.ans",chk_file=run_dir/"data.txt";
+                fil in_file=system_nul,out_file=system_nul,ans_file=run_dir/"data.ans",chk_file=run_dir/"data.txt",tmp_out=fil();
                 if(!i.value()["in"].is_null()) ((fil)(std::string)i.value()["in"]).copyTo((in_file=run_dir/"data.in").path());
                 if(!i.value()["out"].is_null()) ((fil)(std::string)i.value()["out"]).copyTo((out_file=run_dir/"data.out").path());
+                else if(out!=fil()) out_file=run_dir/"data.out",tmp_out=out;
                 // judge
-                judger run_judger(ans,chk,in_file,out_file,ans_file,chk_file);
+                judger run_judger(ans,chk,in_file,out_file,ans_file,chk_file);(&run_judger)->set_name(data_name)->set_out(tmp_out);
                 run_judger.judge();
                 run_judger.print_result("",_chk_name);
                 // print result
@@ -203,19 +223,28 @@ class Command_judge: public App
                 output_chk_file
                 <<"    infile: "<<add_squo(get_file(!i.value()["in"].is_null()?(std::string)i.value()["in"]:system_nul))
                 <<", outfile: "<<add_squo(get_file(!i.value()["out"].is_null()?(std::string)i.value()["out"]:system_nul))<<"\n";
+                if(run_judger.out!=fil()) output_chk_file<<"    "+_out_name+": "<<add_squo(run_judger.out)<<"\n";
                 output_chk_file<<"    result: "<<run_judger.result<<"\n";
                 output_chk_file<<print_type({"    "," time: "," exit_code: ","\n"},{
                     {_ans_name+":",run_judger.time,run_judger.exit_code},
-                    {_chk_name+":",run_judger.chk_runner->time,run_judger.chk_runner->exit_code}});
+                    {_chk_name+":",run_judger.chk_runner->time,run_judger.chk_runner->exit_code},
+                    {_out_name+":",run_judger.out_runner?run_judger.out_runner->time:(tim)-1,run_judger.out_runner?run_judger.out_runner->exit_code:-1}});
                 output_chk_file<<std::string("*")*50;
                 output_chk_file.close();
                 // copy result
                 if(run_judger.result.istrue()) ++ac_sum;
                 if(run_judger.result.isfalse())
                 {
-                    run_dir.copyTo((default_data_dir/(data_name+" - "+get_short_resultname(run_judger.result))).path());
+                    ++output_sum;
+                    pat target_dir=get_file(replace_env(data_file_str,running_path,env_args::data(data_name,run_judger.result,"",runned_sum,output_sum))).path();
+                    ((fil)target_dir.parent()).createDirectories();
+                    if((run_dir/"data.in").exists()) (run_dir/"data.in").copyTo(replace_extension(target_dir,"in").toString());
+                    if((run_dir/"data.out").exists()) (run_dir/"data.out").copyTo(replace_extension(target_dir,"out").toString());
+                    if((run_dir/"data.ans").exists()) (run_dir/"data.ans").copyTo(replace_extension(target_dir,"ans").toString());
+                    if((run_dir/"data.txt").exists()) (run_dir/"data.txt").copyTo(replace_extension(target_dir,"txt").toString());
                 }
-                #undef run_dir
+                ssetenv("runned_sum","");
+                ssetenv("outputed_sum","");
             }
             scout<<ac_sum<<" / "<<runned_sum;
         }